+++ src/tinymesh/core/face.cpp
@@ -18,7 +18,7 @@ bool Face::operator==(const Face &other) const {
     return ret;
 }
 
-Vec3 Face::normal() {
+Vec3 Face::normal() const {
     std::vector<Vec3> vs;
     for (auto it = v_begin(); it != v_end(); ++it) {
         vs.push_back(it->pos());
@@ -37,7 +37,7 @@ Vec3 Face::normal() {
     return normalize(norm);
 }
 
-double Face::area() {
+double Face::area() const {
     std::vector<Vec3> vs;
     for (auto it = v_begin(); it != v_end(); ++it) {
         vs.push_back(it->pos());
@@ -54,7 +54,7 @@ double Face::area() {
     return area;
 }
 
-bool Face::isHole() {
+bool Face::isHole() const {
     // Face is hole if all the vertices are at the boundary.
     for (auto it = this->he_begin(); it != this->he_end(); ++it) {
         if (!it->isBoundary()) {
@@ -64,7 +64,7 @@ bool Face::isHole() {
     return true;
 }
 
-bool Face::isBoundary() {
+bool Face::isBoundary() const {
     for (auto it = this->he_begin(); it != this->he_end(); ++it) {
         if (it->isBoundary()) {
             return true;
@@ -73,7 +73,7 @@ bool Face::isBoundary() {
     return false;
 }
 
-bool Face::isLocked() {
+bool Face::isLocked() const {
     for (auto it = this->v_begin(); it != this->v_end(); ++it) {
         if (it->isLocked()) {
             return true;
@@ -90,6 +90,14 @@ Face::VertexIterator Face::v_end() {
     return Face::VertexIterator(nullptr);
 }
 
+Face::ConstVertexIterator Face::v_begin() const {
+    return Face::ConstVertexIterator(halfedge_);
+}
+
+Face::ConstVertexIterator Face::v_end() const {
+    return Face::ConstVertexIterator(nullptr);
+}
+
 Face::HalfedgeIterator Face::he_begin() {
     return Face::HalfedgeIterator(halfedge_);
 }
@@ -98,12 +106,28 @@ Face::HalfedgeIterator Face::he_end() {
     return Face::HalfedgeIterator(nullptr);
 }
 
+Face::ConstHalfedgeIterator Face::he_begin() const {
+    return Face::ConstHalfedgeIterator(halfedge_);
+}
+
+Face::ConstHalfedgeIterator Face::he_end() const {
+    return Face::ConstHalfedgeIterator(nullptr);
+}
+
 Face::FaceIterator Face::f_begin() {
     return Face::FaceIterator(halfedge_);
 }
 
 Face::FaceIterator Face::f_end() {
     return Face::FaceIterator(nullptr);
+}
+
+Face::ConstFaceIterator Face::f_begin() const {
+    return Face::ConstFaceIterator(halfedge_);
+}
+
+Face::ConstFaceIterator Face::f_end() const {
+    return Face::ConstFaceIterator(nullptr);
 }
 
 // ----------
@@ -149,6 +173,48 @@ Face::VertexIterator Face::VertexIterator::operator++(int) {
 }
 
 // ----------
+// ConstVertexIterator
+// ----------
+
+Face::ConstVertexIterator::ConstVertexIterator(Halfedge *he)
+    : halfedge_{ he }
+    , iter_{ he } {
+}
+
+bool Face::ConstVertexIterator::operator!=(const Face::ConstVertexIterator &it) const {
+    return iter_ != it.iter_;
+}
+
+const Vertex &Face::ConstVertexIterator::operator*() const {
+    return *iter_->src();
+}
+
+const Vertex *Face::ConstVertexIterator::ptr() const {
+    return iter_->src();
+}
+
+const Vertex *Face::ConstVertexIterator::operator->() const {
+    return iter_->src();
+}
+
+Face::ConstVertexIterator &Face::ConstVertexIterator::operator++() {
+    iter_ = iter_->next();
+    if (iter_ == halfedge_) {
+        iter_ = nullptr;
+    }
+    return *this;
+}
+
+Face::ConstVertexIterator Face::ConstVertexIterator::operator++(int) {
+    Halfedge *tmp = iter_;
+    iter_ = iter_->next();
+    if (iter_ == halfedge_) {
+        iter_ = nullptr;
+    }
+    return Face::ConstVertexIterator(tmp);
+}
+
+// ----------
 // HalfedgeIterator
 // ----------
 
@@ -191,6 +257,48 @@ Face::HalfedgeIterator Face::HalfedgeIterator::operator++(int) {
 }
 
 // ----------
+// ConstHalfedgeIterator
+// ----------
+
+Face::ConstHalfedgeIterator::ConstHalfedgeIterator(Halfedge *he)
+    : halfedge_{ he }
+    , iter_{ he } {
+}
+
+bool Face::ConstHalfedgeIterator::operator!=(const Face::ConstHalfedgeIterator &it) const {
+    return iter_ != it.iter_;
+}
+
+const Halfedge &Face::ConstHalfedgeIterator::operator*() const {
+    return *iter_;
+}
+
+const Halfedge *Face::ConstHalfedgeIterator::ptr() const {
+    return iter_;
+}
+
+const Halfedge *Face::ConstHalfedgeIterator::operator->() const {
+    return iter_;
+}
+
+Face::ConstHalfedgeIterator &Face::ConstHalfedgeIterator::operator++() {
+    iter_ = iter_->next();
+    if (iter_ == halfedge_) {
+        iter_ = nullptr;
+    }
+    return *this;
+}
+
+Face::ConstHalfedgeIterator Face::ConstHalfedgeIterator::operator++(int) {
+    Halfedge *tmp = iter_;
+    iter_ = iter_->next();
+    if (iter_ == halfedge_) {
+        iter_ = nullptr;
+    }
+    return Face::ConstHalfedgeIterator(tmp);
+}
+
+// ----------
 // FaceIterator
 // ----------
 
@@ -230,6 +338,48 @@ Face::FaceIterator Face::FaceIterator::operator++(int) {
         iter_ = nullptr;
     }
     return Face::FaceIterator(tmp);
+}
+
+// ----------
+// ConstFaceIterator
+// ----------
+
+Face::ConstFaceIterator::ConstFaceIterator(Halfedge *he)
+    : halfedge_{ he }
+    , iter_{ he } {
+}
+
+bool Face::ConstFaceIterator::operator!=(const Face::ConstFaceIterator &it) const {
+    return iter_ != it.iter_;
+}
+
+const Face &Face::ConstFaceIterator::operator*() const {
+    return *iter_->rev()->face();
+}
+
+const Face *Face::ConstFaceIterator::ptr() const {
+    return iter_->rev()->face();
+}
+
+const Face *Face::ConstFaceIterator::operator->() const {
+    return iter_->rev()->face();
+}
+
+Face::ConstFaceIterator &Face::ConstFaceIterator::operator++() {
+    iter_ = iter_->next();
+    if (iter_ == halfedge_) {
+        iter_ = nullptr;
+    }
+    return *this;
+}
+
+Face::ConstFaceIterator Face::ConstFaceIterator::operator++(int) {
+    Halfedge *tmp = iter_;
+    iter_ = iter_->next();
+    if (iter_ == halfedge_) {
+        iter_ = nullptr;
+    }
+    return Face::ConstFaceIterator(tmp);
 }
 
 }  // namespace tinymesh